Moved KonanTarget to shared. Placed program suffix into KonanTarget. A little simplified the handling of output name. Made samples work with the new TargetManager location.
