+++ backend.native/build.gradle
@@ -57,7 +57,10 @@ sourceSets {
           srcDir 'compiler/ir/backend.native/src/'
           srcDir 'build/renamed/source/proto/compiler/java'
       }
-      kotlin.srcDir 'compiler/ir/backend.native/src/'
+      kotlin {
+          srcDir 'compiler/ir/backend.native/src/'
+          srcDir "${rootProject.projectDir}/shared/src/main/kotlin"
+      }
     }
     cli_bc {
       java.srcDir 'cli.bc/src'
@@ -130,6 +133,7 @@ dependencies {
    compilerCompile kotlinCompilerModule
    compilerCompile kotlinNativeInterop['llvm'].configuration
    compilerCompile kotlinNativeInterop['hash'].configuration
+   compilerCompile project(':shared')
 
    cli_bcCompile "org.jetbrains.kotlin:kotlin-runtime:$kotlin_version"
    cli_bcCompile sourceSets.compiler.output
