+++ backend.native/cli.bc/src/org/jetbrains/kotlin/cli/bc/K2Native.kt
@@ -18,10 +18,7 @@ package org.jetbrains.kotlin.cli.bc
 
 import com.intellij.openapi.Disposable
 import org.jetbrains.kotlin.backend.konan.*
-import org.jetbrains.kotlin.backend.konan.CompilerOutputKind.*
 import org.jetbrains.kotlin.backend.konan.util.profile
-import org.jetbrains.kotlin.backend.konan.util.suffixIfNot
-import org.jetbrains.kotlin.backend.konan.util.removeSuffixIfPresent
 import org.jetbrains.kotlin.cli.common.CLICompiler
 import org.jetbrains.kotlin.cli.common.CLITool
 import org.jetbrains.kotlin.cli.common.ExitCode
@@ -31,6 +28,7 @@ import org.jetbrains.kotlin.config.CommonConfigurationKeys
 import org.jetbrains.kotlin.config.CompilerConfiguration
 import org.jetbrains.kotlin.config.Services
 import org.jetbrains.kotlin.config.addKotlinSourceRoots
+import org.jetbrains.kotlin.konan.target.*
 import java.util.*
 import kotlin.reflect.KFunction
 
@@ -109,18 +107,10 @@ class K2Native : CLICompiler<K2NativeCompilerArguments>() {
                 // TODO: Collect all the explicit file names into an object
                 // and teach the compiler to work with temporaries and -save-temps.
 
+                arguments.outputName ?.let { put(OUTPUT, it) } 
                 val outputKind = CompilerOutputKind.valueOf(
                     (arguments.produce ?: "program").toUpperCase())
-
                 put(PRODUCE, outputKind)
-                val suffix = outputKind.suffix
-                val output = arguments.outputFile?.removeSuffixIfPresent(suffix) 
-                    ?: outputKind.name.toLowerCase()
-                put(OUTPUT_NAME, output)
-                put(OUTPUT_FILE, output.suffixIfNot(outputKind.suffix))
-
-                // This is a decision we could change
-                put(CommonConfigurationKeys.MODULE_NAME, output)
                 put(ABI_VERSION, 1)
 
                 arguments.mainPackage ?.let{ put(ENTRY, it) }
