+++ backend.native/cli.bc/src/org/jetbrains/kotlin/cli/bc/K2NativeCompilerArguments.kt
@@ -54,8 +54,8 @@ class K2NativeCompilerArguments : CommonCompilerArguments() {
     @field:Argument(value = "-opt", description = "Enable optimizations during compilation")
     @JvmField var optimization: Boolean = false
 
-    @field:Argument(value = "-output", shortName = "-o", valueDescription = "<path>", description = "Output file path")
-    @JvmField var outputFile: String? = null
+    @field:Argument(value = "-output", shortName = "-o", valueDescription = "<name>", description = "Output name")
+    @JvmField var outputName: String? = null
 
     @field:Argument(value = "-entry", shortName = "-e", valueDescription = "<name>", description = "Qualified entry point name")
     @JvmField var mainPackage: String? = null
